added error handling in case of connection lost with server. when client loses his connetion displays messagebox and redisplay window where user can enter nickname and ip address again
added loop-like behavior for client - if connection to the server doesnt work
user will now get messageBox if option to retrying to connect or canceling and
maybe try other ip address

added console mode for debugging

added user32.dll for qmake

added c++11 casts

added object file to git ignore, and dependencies folder that contains qt dlls for client

+++ qt-client/chatwindow.cpp
@@ -108,7 +108,7 @@ void ChatWindow::sendButtonClicked()
 
     mlp.fillFrame('1',"",msg.toLocal8Bit().constData());
     string sendbuff = mlp.packFrame();
-    send(this->getSock(), sendbuff.c_str(), sendbuff.size()+1,0);
+    send(this->getSock(), sendbuff.c_str(), static_cast<int>(sendbuff.size()+1),0);
 }
 
 void ChatWindow::adjustLabelSlot()
@@ -126,5 +126,5 @@ void ChatWindow::sendMSG()
 
     mlp.fillFrame('1',"",msg.toLocal8Bit().constData());
     string sendbuff = mlp.packFrame();
-    send(this->getSock(), sendbuff.c_str(), sendbuff.size()+1,0);
+    send(this->getSock(), sendbuff.c_str(), static_cast<int>(sendbuff.size()+1),0);
 }
